fix: additional script bugs found during testing
- streamlines.py: fix dump_prefix='flow' -> 'grid' in create_snapshot
- streamlines.py: fill NaN gaps with nearest-neighbour for streamplot
- animate_particles.py: cast type column to int for color indexing
- grid_density_heatmap.py, grid_pressure_heatmap.py: resolve folder_path
  and in.ampt relative to _REPO_ROOT
- plot_drag.py, analyze_multi_altitude_temp.py, analyze_multi_altitude_origin.py:
  resolve all dumps/outputs/in.ampt paths relative to _REPO_ROOT

All scripts now work from any working directory.

+++ scripts/animate_particles.py
@@ -64,7 +64,7 @@ def update(frame_idx):
     x = dff['x'].values
     y = dff['y'].values
     z = dff['z'].values
-    t = dff['type'].astype(int).map(type_to_color).fillna(0).values
+    t = dff['type'].astype(int).map(type_to_color).fillna(0).astype(int).values
     c = [colors_cycle[i % len(colors_cycle)] for i in t]
     scat._offsets3d = (x, y, z)
     scat.set_color(c)
